Big releease - support ideally for clang by enabling a ton of compiler flags (more strict)

+++ elements/001-hydrogen/hydrogen/src/api/system/config/config.c
@@ -93,8 +93,8 @@ enum MHD_Result handle_system_config_request(struct MHD_Connection *connection,
     // Add timing information
     struct timespec end_time;
     clock_gettime(CLOCK_MONOTONIC, &end_time);
-    double processing_time = (end_time.tv_sec - start_time.tv_sec) * 1000.0 + 
-                           (end_time.tv_nsec - start_time.tv_nsec) / 1000000.0;
+    double processing_time = ((double)(end_time.tv_sec - start_time.tv_sec)) * 1000.0 + 
+                           ((double)(end_time.tv_nsec - start_time.tv_nsec)) / 1000000.0;
     
     json_t *timing = json_object();
     json_object_set_new(timing, "processing_time_ms", json_real(processing_time));
